// src/types/type.ts
export enum Routes {
  onBoardingStack = 'onBoardingStack',
  welcomeScreen = 'Welcome',
  enterCompanyIdScreen = 'Enter Company ID',
  mainScreenStack = 'mainScreenStack',
  mainScreen = 'Main',
  voiceBotScreen = 'Voice Bot',
  settingsStack = 'settingsStack',
  settingsScreen = 'Settings',
  setCompanyIdScreen = 'Set Company Id',
  pickVoiceScreen = 'Pick Voice',
}